% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuRacy.R
\name{accuRacy}
\alias{accuRacy}
\title{Calculate Classification Accuracy Metrics}
\usage{
accuRacy(classified_img, training_shapefile, validation_shapefile)
}
\arguments{
\item{classified_img}{A Raster* object representing the classified image.}

\item{training_shapefile}{Path to the training shapefile.}

\item{validation_shapefile}{Path to the validation shapefile.}
}
\value{
A list containing the accuracy matrix, user's accuracy, producer's accuracy, and overall accuracy.
}
\description{
This function calculates classification accuracy metrics, including user's accuracy (UA),
producer's accuracy (PA), and overall accuracy (OA), based on a classified image and
reference shapefiles for training and validation.
}
\examples{
\dontrun{
# Train the Random Forest model
rf_model <- tRain(smp)
# Obtain classification result
classification_file <- Ranger(img, rf_model)
# Accuracy assessment
accuracy_metrics <- accuRacy(classification_RF.tif, TrainingPoints.shp, ValidationPoints.shp)
print(accuracy_metrics)
}
}
