% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extRact.R
\name{extRact}
\alias{extRact}
\title{Extract Raster Values Based on Shapefile}
\usage{
extRact(img, shp)
}
\arguments{
\item{img}{Path to the Raster file}

\item{shp}{Path to the shapefile.}
}
\value{
A data frame containing the extracted raster values.
}
\description{
This function extracts raster values based on the polygons defined in a shapefile.
It assigns factor levels from the shapefile and removes the first column (ID)
from the resulting data frame.
}
\examples{
\dontrun{
extracted_data <- extRact(img, shp)
}
}
