% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaData.R
\name{loaData}
\alias{loaData}
\title{Load Raster and Shapefile Data}
\usage{
loaData(raster_path, shapefile_path)
}
\arguments{
\item{raster_path}{Path to the raster file.}

\item{shapefile_path}{Path to the shapefile.}
}
\value{
A list containing the loaded raster and shapefile data.
}
\description{
This function loads raster and shapefile data from specified paths. It prints
information about the loaded data, including the raster data, shapefile data,
and labels of the shapefile.
}
\examples{
\dontrun{
# Load raster and shapefile data from "inst/extdata" folder
raster_path<- "AOI.tif"
shapefile_path <- "TrainingPoints.shp"
loaData(raster_path, shapefile_path)
}
}
