% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ranger.R
\name{Ranger}
\alias{Ranger}
\title{Predict Image Data with Random Forest Model}
\usage{
Ranger(img, rfmodel, filename = "classification_RF.tif", overwrite = TRUE)
}
\arguments{
\item{img}{A Raster* image representing the image data to be classified.}

\item{rfmodel}{A trained Random Forest model object.}

\item{filename}{Optional. The filename for the resulting classified image. Defaults to "classification_RF.tif".}

\item{overwrite}{Logical. Whether to overwrite an existing file with the same name as 'filename'. Defaults to TRUE}
}
\value{
A RasterLayer object representing the classified image.
}
\description{
This function predicts forest type classification for image data using a trained model.
}
\examples{
\dontrun{
classified_image <- Ranger(AOI.tif, rf_model, filename = "classification_RF.tif")
}
}
