% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rename.R
\name{Rename}
\alias{Rename}
\title{Rename Bands in Raster Object}
\usage{
Rename(img)
}
\arguments{
\item{img}{A Raster* image.}
}
\value{
The preprocessed image data with dynamically assigned band names
}
\description{
This function preprocesses the image data by assigning band names dynamically
based on the number of bands in the image.
}
\examples{
\dontrun{
renamed_img <- Rename(img)
print(names(renamed_img))
}
}
